requirements.txt
# Core dependencies with compatible versions
chromadb==0.4.22
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-chroma==0.1.0
sentence-transformers==2.2.2
groq==0.4.1
numpy==1.24.3
torch==2.1.2
torchvision==0.16.2
huggingface-hub==0.17.3

# Additional dependencies
python-dotenv==1.0.0
colorama==0.4.6